requirements.txt
requests
pygame
openai==0.27.0
python-dotenv
python-telegram-bot
opencv-python